Added (u32) cast to Xil_In/Out32 addr args to fix pointer/integer compiler warnings
git-svn-id: http://warpproject.org/svn/WARP@6023 dd934ca3-7c2e-414f-9084-136bfdd87501

+++ ReferenceDesigns/w3_802.11/c-dev/wlan_mac_high_framework/wlan_exp_node.c
@@ -2774,7 +2774,7 @@ int process_node_cmd(int socket_index, void * from, cmd_resp * command, cmd_resp
                         resp_hdr->num_args = resp_index;
 
                         for (mem_idx = 0; mem_idx < mem_length; mem_idx++) {
-                            resp_args_32[resp_index + mem_idx] = Xil_Ntohl(Xil_In32((void*)(mem_addr) + mem_idx*sizeof(u32)));
+                            resp_args_32[resp_index + mem_idx] = Xil_Ntohl(Xil_In32((u32)((void*)(mem_addr) + mem_idx*sizeof(u32))));
                         }
 
                         // Update response header with payload length
